backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "payments-backend"
version = "0.1.0"
description = "Clean architecture FastAPI backend for contractor payments"
authors = [
    {name = "vr-ski", email = "166657596+vr-ski@users.noreply.github.com"}
]
requires-python = ">=3.14"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.30",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.1",
    "alembic>=1.13.1",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.21"
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.0.1",
    "pytest>=8.2.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.5",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "black>=24.4.0"
]

[tool.setuptools.packages.find]
where = ["app"]

[tool.ruff]
line-length = 88
target-version = "py314"
lint.select = ["E", "F", "I", "UP", "B", "C4", "SIM", "ARG", "PTH", "RUF"]
lint.ignore = []

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
files = ["app"]
